Added ForgotPassword and api for orders; Fixed CartScreen and strings resource

+++ app/src/main/java/com/example/matule/ui/presentation/feature/cart/ui/CartScreen.kt
@@ -1,6 +1,18 @@
 package com.example.matule.ui.presentation.feature.cart.ui
 
+import android.app.AlertDialog
+import android.content.Context
+import android.widget.Toast
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.scaleIn
+import androidx.compose.animation.scaleOut
+import androidx.compose.animation.shrinkVertically
 import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -13,23 +25,30 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Icon
+import androidx.compose.material3.SnackbarDuration
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.blur
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.PathEffect
-import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
@@ -39,17 +58,25 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import com.example.domain.ui.presentation.feature.cart.model.CartItem
-import com.example.domain.ui.presentation.feature.cart.model.CartState
+import com.example.domain.ui.presentation.feature.orders.model.Address
+import com.example.domain.ui.presentation.feature.orders.model.ContactInfo
+import com.example.domain.ui.presentation.feature.orders.model.CreateOrderRequest
+import com.example.domain.ui.presentation.feature.orders.model.Order
+import com.example.domain.ui.presentation.feature.orders.model.PaymentMethod
+import com.example.domain.ui.presentation.feature.profile.model.UserProfile
 import com.example.matule.R
+import com.example.matule.ui.presentation.approuts.AppRouts
 import com.example.matule.ui.presentation.feature.cart.viewmodel.CartScreenContract
 import com.example.matule.ui.presentation.feature.cart.viewmodel.CartScreenViewModel
 import com.example.matule.ui.presentation.shared.buttons.CustomButton
 import com.example.matule.ui.presentation.shared.cart.CartItem
+import com.example.matule.ui.presentation.shared.cart.InformationSection
 import com.example.matule.ui.presentation.shared.header.CustomHeader
 import com.example.matule.ui.presentation.shared.screen.EmptyScreen
 import com.example.matule.ui.presentation.shared.screen.MainLoadingScreen
 import com.example.matule.ui.presentation.theme.Colors
 import com.example.matule.ui.presentation.theme.MatuleTypography
+import kotlinx.coroutines.launch
 import java.text.SimpleDateFormat
 import java.util.Locale
 import java.util.TimeZone
@@ -58,7 +85,7 @@ private interface CartScreenCallback{
     fun onBack() {}
     fun openDetailScreen(id: Long) {}
     fun deleteProduct(id: Long) {}
-    fun checkoutScreen() {}
+    fun checkoutScreen(request: CreateOrderRequest) {}
     fun updateQuality(cartItemId: Long, newQuantity: Int) {}
 }
 
@@ -68,8 +95,39 @@ fun CartScreen(
     navController: NavHostController
 ) {
     val state by vm.state.collectAsState()
+    val profile by vm.profile.collectAsState()
+    val effect = vm.effect.collectAsState(initial = null)
+    val context = LocalContext.current
+    val snackbarHostState = remember { SnackbarHostState() }
+    val coroutineScope = rememberCoroutineScope()
 
     var visibleCheckout by remember { mutableStateOf(false) }
+    var visibleSnackbar by remember { mutableStateOf(false) }
+
+    LaunchedEffect(effect.value) {
+        effect.value?.let { effect ->
+            when (effect) {
+                is CartScreenContract.Effect.OrderCreated -> {
+                    showOrderSuccessDialog(context, effect.order, effect.message)
+                }
+                is CartScreenContract.Effect.ShowToast -> {
+                    Toast.makeText(context, effect.message, Toast.LENGTH_SHORT).show()
+                }
+                is CartScreenContract.Effect.ShowSnackbar -> {
+                    coroutineScope.launch {
+                        val snackbarResult = snackbarHostState.showSnackbar(
+                            message = effect.message,
+                            duration = SnackbarDuration.Short
+                        )
+                    }
+                }
+            }
+        }
+    }
+
+    LaunchedEffect(Unit) {
+        vm.handleEvent(CartScreenContract.Event.LoadCart)
+    }
 
     val callback = object : CartScreenCallback{
         override fun onBack() {
@@ -80,8 +138,9 @@ fun CartScreen(
             }
         }
 
-        override fun checkoutScreen() {
-            visibleCheckout = true
+        override fun checkoutScreen(request: CreateOrderRequest) {
+           vm.handleEvent(CartScreenContract.Event.CreateOrder(request))
+           visibleSnackbar = true
         }
 
         override fun deleteProduct(id: Long) {
@@ -89,6 +148,7 @@ fun CartScreen(
         }
 
         override fun openDetailScreen(id: Long) {
+            navController.navigate("${AppRouts.DETAILS}/$id")
         }
 
         override fun updateQuality(cartItemId: Long, newQuantity: Int) {
@@ -98,18 +158,37 @@ fun CartScreen(
 
     CartContent(
         callback = callback,
-        state = state
+        state = state,
+        profile = profile,
+        visibleCheckout = visibleCheckout,
+        onShowCheckoutForm = { visibleCheckout = true },
+        visibleSnackBar = visibleSnackbar
     )
+
+    AnimatedVisibility(
+       visible = visibleSnackbar,
+       enter = fadeIn() + scaleIn(tween(150)),
+       exit = fadeOut() + scaleOut(tween(150))
+    ) {
+        CardCongratulations {
+            navController.navigate(AppRouts.MAIN)
+        }
+    }
 }
 
 @Composable
 private fun CartContent(
+    visibleCheckout: Boolean,
+    visibleSnackBar: Boolean,
+    profile: UserProfile,
+    onShowCheckoutForm: () -> Unit,
     state: CartScreenContract.State,
     callback: CartScreenCallback,
 ) {
     Column(
         modifier = Modifier
-            .background(color = Colors.background)
+            .background(color = if (visibleSnackBar) Colors.background.copy(.2f) else Colors.background)
+            .blur(if (visibleSnackBar) 4.dp else 0.dp)
             .fillMaxSize()
     ) {
         CustomHeader(
@@ -118,12 +197,11 @@ private fun CartContent(
             modifier = Modifier.padding(horizontal = 20.dp),
             onBack = callback::onBack
         )
-        Spacer(modifier = Modifier.height(16.dp))
+        Spacer(modifier = Modifier.height(if (visibleCheckout) 46.dp else 16.dp))
         when(state) {
             is CartScreenContract.State.Loaded -> {
                 Cart(
                     state = state,
-                    openCheckoutScreen = callback::checkoutScreen,
                     openDetailScreen = {
                         callback.openDetailScreen(it)
                     },
@@ -134,19 +212,22 @@ private fun CartContent(
                         callback.updateQuality(id, quality + 1)
                     },
                     onMinusQuantity = { id, quality ->
-                        callback.updateQuality(id, quality - 1)
-                    }
+                        if (quality > 1) {
+                            callback.updateQuality(id, quality - 1)
+                        }
+                    },
+                    profile = profile,
+                    onShowCheckoutForm = onShowCheckoutForm,
+                    onCreateOrder = {
+                        callback.checkoutScreen(it)
+                    },
+                    visibleCheckout = visibleCheckout
                 )
             }
             is CartScreenContract.State.Loading -> {
                 MainLoadingScreen()
             }
-            is CartScreenContract.State.Empty -> {
-                EmptyScreen(
-                    icon = R.drawable.ic_cart,
-                    emptyText = R.string.empty_cart
-                )
-            }
+            else -> {}
         }
     }
 }
@@ -154,28 +235,89 @@ private fun CartContent(
 @Composable
 fun Cart(
     state: CartScreenContract.State.Loaded,
-    openCheckoutScreen: () -> Unit,
+    profile: UserProfile,
+    visibleCheckout: Boolean,
     openDetailScreen: (Long) -> Unit,
     onDelete: (Long) -> Unit,
+    onShowCheckoutForm: () -> Unit,
+    onCreateOrder: (CreateOrderRequest) -> Unit,
     onPlusQuantity: (Long, Int) -> Unit,
     onMinusQuantity: (Long, Int) -> Unit
 ) {
+
     Column(
         modifier = Modifier.fillMaxSize(),
         verticalArrangement = Arrangement.SpaceBetween
     ) {
-        Content(
-            cartProduct = state.items,
-            openDetailScreen = openDetailScreen,
-            onDelete = onDelete,
-            onPlusQuantity = onPlusQuantity,
-            onMinusQuantity = onMinusQuantity
-        )
+        AnimatedVisibility(
+            visible = !visibleCheckout,
+            enter = expandVertically(
+                animationSpec = tween(700),
+                expandFrom = Alignment.Top
+            ) + fadeIn(animationSpec = tween(700)),
+            exit = shrinkVertically(
+                animationSpec = tween(700),
+                shrinkTowards = Alignment.Top
+            ) + fadeOut(animationSpec = tween(700))
+        ) {
+            Content(
+                cartProduct = state.items,
+                openDetailScreen = openDetailScreen,
+                onDelete = onDelete,
+                onPlusQuantity = onPlusQuantity,
+                onMinusQuantity = onMinusQuantity
+            )
+        }
+        AnimatedVisibility(
+            visible = visibleCheckout,
+            enter = expandVertically(
+                animationSpec = tween(700),
+                expandFrom = Alignment.Top
+            ) + fadeIn(animationSpec = tween(700)),
+            exit = shrinkVertically(
+                animationSpec = tween(700),
+                shrinkTowards = Alignment.Top
+            ) + fadeOut(animationSpec = tween(700))
+        ) {
+            InformationSection(
+                address = profile.address,
+                city = profile.city,
+                country = profile.country,
+                postalCode = profile.postalCode,
+                phone = profile.phone,
+                visibleEndIcon = true,
+                email = profile.email
+            )
+        }
+
         BottomBarCart(
             subtotal = state.subtotal,
             delivery = state.delivery,
             totalCost = state.totalPrice,
-            openCheckoutScreen = openCheckoutScreen
+            openCheckoutScreen = {
+                if (!visibleCheckout) {
+                    onShowCheckoutForm()
+                } else {
+                    val request = CreateOrderRequest(
+                        contactInfo = ContactInfo(
+                            firstName = profile.firstName,
+                            lastName = profile.lastName,
+                            phone = profile.phone ?: "",
+                            email = profile.email
+                        ),
+                        address = Address(
+                            country = profile.country ?: "",
+                            city = profile.city ?: "",
+                            street = profile.address ?: "",
+                            postalCode = profile.postalCode,
+                            apartment = ""
+                        ),
+                        paymentMethod = PaymentMethod.fromString("Card"),
+                    )
+                    onCreateOrder(request)
+                }
+            },
+            visibleCheckout = visibleCheckout
         )
     }
 }
@@ -202,27 +344,38 @@ fun Content(
             fontFamily = FontFamily(Font(R.font.new_peninim_mt_inclined_2))
         )
         Spacer(modifier = Modifier.height(16.dp))
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .verticalScroll(rememberScrollState()),
-            verticalArrangement = Arrangement.spacedBy(14.dp)
-        ) {
-            cartProduct.forEach { cart ->
+        if (cartProduct.isNotEmpty()) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .verticalScroll(rememberScrollState()),
+                verticalArrangement = Arrangement.spacedBy(14.dp)
+            ) {
+                cartProduct.forEach { cart ->
 
-                val timeAgoInMinutes = calculateMinutesAgo(cart.createdAt)
-                CartItem(
-                    timeAgo = timeAgoInMinutes,
-                    quantity = cart.quantity,
-                    photoProduct = cart.product?.images?.first() ?: "",
-                    nameProduct = cart.product?.name ?: "",
-                    price = cart.product?.price ?: 0.0,
-                    openDetailScreen = { openDetailScreen(cart.product?.id ?: 0) },
-                    onDelete = { onDelete(cart.id) },
-                    onMinusQuantity = { onMinusQuantity(cart.id, cart.quantity) },
-                    onPlusQuantity = { onPlusQuantity(cart.id, cart.quantity) }
-                )
+                    val timeAgoInMinutes = calculateMinutesAgo(cart.createdAt)
+                    CartItem(
+                        timeAgo = timeAgoInMinutes,
+                        quantity = cart.quantity,
+                        photoProduct = cart.product?.images?.first() ?: "",
+                        nameProduct = cart.product?.name ?: "",
+                        price = cart.product?.price ?: 0.0,
+                        openDetailScreen = { openDetailScreen(cart.product?.id ?: 0) },
+                        onDelete = { onDelete(cart.id) },
+                        onMinusQuantity = { onMinusQuantity(cart.id, cart.quantity) },
+                        onPlusQuantity = { onPlusQuantity(cart.id, cart.quantity) }
+                    )
+
+                }
             }
+        } else {
+            EmptyScreen(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 100.dp),
+                icon = R.drawable.ic_cart,
+                emptyText = R.string.empty_cart
+            )
         }
     }
 }
@@ -230,6 +383,7 @@ fun Content(
 @Composable
 fun BottomBarCart(
     subtotal: Double,
+    visibleCheckout: Boolean,
     delivery: Double,
     totalCost: Double,
     openCheckoutScreen: () -> Unit
@@ -265,7 +419,7 @@ fun BottomBarCart(
         )
         Spacer(modifier = Modifier.height(32.dp))
         CustomButton(
-            text = R.string.btn_checkout,
+            text = if (visibleCheckout) R.string.btn_confirm else R.string.btn_checkout,
             onClick = openCheckoutScreen,
             modifier = Modifier.fillMaxWidth()
         )
@@ -322,6 +476,60 @@ fun DashedLine(
     }
 }
 
+@Composable
+fun CardCongratulations(
+    openMainScreen: () -> Unit
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize(),
+        contentAlignment = Alignment.BottomCenter
+    ) {
+        Column(
+            modifier = Modifier
+                .padding(
+                    start = 20.dp, end = 20.dp, bottom = 153.dp
+                )
+                .background(color = Colors.block, RoundedCornerShape(20.dp))
+                .padding(horizontal = 40.dp, vertical = 49.dp),
+            verticalArrangement = Arrangement.spacedBy(30.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Column(
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.spacedBy(24.dp)
+            ) {
+                Box(
+                    modifier = Modifier
+                        .background(color = Color(0xFFDFEFFF), shape = CircleShape)
+                ) {
+                    Image(
+                        painter = painterResource(R.drawable.congratulations),
+                        contentDescription = null,
+                        modifier = Modifier
+                            .padding(24.dp)
+                            .size(86.dp)
+                    )
+                }
+                Text(
+                    text = stringResource(R.string.checkout_popup),
+                    color = Colors.text,
+                    fontSize = 20.sp,
+                    lineHeight = 28.sp,
+                    textAlign = TextAlign.Center,
+                    fontWeight = FontWeight.Normal,
+                    fontFamily = FontFamily(Font(R.font.new_peninim_mt_inclined_2))
+                )
+            }
+            CustomButton(
+                modifier = Modifier.fillMaxWidth(),
+                text = R.string.btn_back_to_shopping,
+                onClick = openMainScreen
+            )
+        }
+    }
+}
+
 fun calculateMinutesAgo(createdAt: String): Int {
     return try {
         val dateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", Locale.getDefault())
@@ -336,7 +544,21 @@ fun calculateMinutesAgo(createdAt: String): Int {
         } else {
             0
         }
-    } catch (e: Exception) {
+    } catch (_: Exception) {
         0
     }
+}
+
+private fun showOrderSuccessDialog(context: Context, order: Order, message: String) {
+    AlertDialog.Builder(context)
+        .setTitle("Заказ создан!")
+        .setMessage("$message\nНомер заказа: #${order.orderNumber}\nСумма: ${order.total.formatPrice()}")
+        .setPositiveButton("ОК") { dialog, _ ->
+            dialog.dismiss()
+        }
+        .show()
+}
+
+fun Double.formatPrice(): String {
+    return String.format("%.2f ₽", this)
 }
